+++ src/commands/package/bundle/install.ts
@@ -136,6 +136,7 @@ export class PackageBundlesInstall extends SfCommand<BundleSObjects.PkgBundleVer
       this.spinner.stop();
     }
 
+    // eslint-disable-next-line @typescript-eslint/switch-exhaustiveness-check
     switch (result.InstallStatus) {
       case BundleSObjects.PkgBundleVersionInstallReqStatus.error: {
         const errorText =
@@ -155,7 +156,7 @@ export class PackageBundlesInstall extends SfCommand<BundleSObjects.PkgBundleVer
       default:
         this.log(
           messages.getMessage('bundleInstallInProgress', [
-            camelCaseToTitleCase(result.InstallStatus as string),
+            camelCaseToTitleCase(result.InstallStatus),
             result.Id,
             targetOrg.getUsername() ?? '',
           ])