+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## [3.0.6](https://github.com/salesforcecli/plugin-packaging/compare/3.0.5...3.0.6) (2026-08-20)
+
+### Bug Fixes
+
+- rejigger eslint.config.mjs (W-23928423) ([64eaa4c](https://github.com/salesforcecli/plugin-packaging/commit/64eaa4cb72f16f659504bbc3ffc4018a4a808fbf))
+
+## [3.0.5](https://github.com/salesforcecli/plugin-packaging/compare/3.0.4...3.0.5) (2026-08-09)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/core from 9.0.0 to 9.1.0 ([b9dd777](https://github.com/salesforcecli/plugin-packaging/commit/b9dd777c5a33265b23573c7068847dad4646f0c7))
+
+## [3.0.4](https://github.com/salesforcecli/plugin-packaging/compare/3.0.3...3.0.4) (2026-08-08)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/ts-types from 3.0.0 to 3.0.1 ([cba5ff6](https://github.com/salesforcecli/plugin-packaging/commit/cba5ff6b9eb0a2fc58276bd398efa1819dcefba8))
+
+## [3.0.3](https://github.com/salesforcecli/plugin-packaging/compare/3.0.2...3.0.3) (2026-08-08)
+
+### Bug Fixes
+
+- **deps:** bump @oclif/core from 4.13.2 to 4.13.3 ([8f4101e](https://github.com/salesforcecli/plugin-packaging/commit/8f4101e6df3df57daf1f5c2c62b263664079fb9a))
+
+## [3.0.2](https://github.com/salesforcecli/plugin-packaging/compare/3.0.1...3.0.2) (2026-08-08)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/packaging from 5.0.0 to 5.0.4 ([c41ce45](https://github.com/salesforcecli/plugin-packaging/commit/c41ce459937f9d0b483e2ea51abffc422a385a7c))
+
+## [3.0.1](https://github.com/salesforcecli/plugin-packaging/compare/3.0.0...3.0.1) (2026-08-06)
+
+### Bug Fixes
+
+- **deps:** bump fast-uri from 3.1.4 to 3.1.5 ([ee53b59](https://github.com/salesforcecli/plugin-packaging/commit/ee53b59582700ad1feb5af42089b9b1ee454a93e))
+
 # [3.0.0](https://github.com/salesforcecli/plugin-packaging/compare/2.30.6...3.0.0) (2026-07-30)
 
 - feat!: require Node >=22, drop EOL Node versions @W-23480655@ ([dfc1e1d](https://github.com/salesforcecli/plugin-packaging/commit/dfc1e1daabe044695ad2c90864898c67fda265d8))